Updating TurningCircle
